Update DeckOfCards.cs
Removed `init()` in favor of a constructor. Whenever a `new Deck()` is initialized, this will ensure the object is properly formed for the creator. Much better than `init()` because `init()` requires knowledge of the internals of the class `Deck` to know that `init()` must be
called before `draw` or `shuffle`. 

There's a concept of 'falling into a pit of success', where good design intent helps developers new and old (or even yourself coming back to code after a while) be successful almost by accident.